requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.9.2
python-multipart==0.0.6

# AI & LLM
anthropic==0.7.8
openai==1.3.7
langchain>=0.1.0
langchain-community>=0.0.6

# Vector Database
chromadb==0.5.18
sentence-transformers==2.2.2

# HTTP & API
httpx>=0.27.0,<1
requests==2.31.0

# Data Processing
numpy==2.1.3
pandas==2.2.2

# Environment & Config
python-dotenv==1.0.0

# Utilities
python-dateutil==2.8.2

# Testing (Optional)
pytest==7.4.3
pytest-asyncio==0.21.1

# Development (Optional)
black==23.11.0
flake8==6.1.0